Move NamespacedPage up a directory, fix all imports to use @console/shared
