+++ frontend/packages/console-shared/src/components/NamespacedPage.scss
