+++ frontend/packages/console-shared/src/components/NamespacedPage.tsx
