+++ frontend/packages/console-shared/src/components/projects/index.ts
@@ -1,0 +1,2 @@
+export { default as CreateProjectListPage } from './CreateProjectListPage';
+export { default as ProjectListPage } from './ProjectListPage';
