+++ frontend/packages/dev-console/src/components/SearchPage.tsx
@@ -4,8 +4,7 @@ import { useTranslation, Trans } from 'react-i18next';
 import { match as RMatch } from 'react-router';
 import { NamespacedPageVariants } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 import { withStartGuide } from '@console/internal/components/start-guide';
-import CreateProjectListPage from '@console/shared/src/components/projects/CreateProjectListPage';
-import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
+import { CreateProjectListPage, NamespacedPage } from '@console/shared';
 import { SearchPage } from '../../../../public/components/search';
 
 export interface SearchPageProps {
