+++ frontend/packages/dev-console/src/components/import/ImportPage.tsx
@@ -6,8 +6,8 @@ import { RouteComponentProps } from 'react-router-dom';
 import { NamespacedPageVariants } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 import { PageHeading, Firehose, FirehoseResource } from '@console/internal/components/utils';
 import { ImageStreamModel, ProjectModel } from '@console/internal/models';
+import { NamespacedPage } from '@console/shared';
 import DevPreviewBadge from '@console/shared/src/components/badges/DevPreviewBadge';
-import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
 import { QUERY_PROPERTIES } from '../../const';
 import QueryFocusApplication from '../QueryFocusApplication';
 import { ImportTypes, ImportData } from './import-types';
