+++ frontend/packages/helm-plugin/src/components/details-page/__tests__/HelmReleaseDetailsPage.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { shallow, ShallowWrapper } from 'enzyme';
-import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
+import { NamespacedPage } from '@console/shared';
 import HelmReleaseDetails from '../HelmReleaseDetails';
 import HelmReleaseDetailsPage from '../HelmReleaseDetailsPage';
