+++ frontend/packages/helm-plugin/src/components/forms/ProjectHelmChartRepository/CreateProjectHelmChartRepositoryPage.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import Helmet from 'react-helmet';
 import { match as RMatch } from 'react-router-dom';
 import { NamespacedPageVariants } from '@console/dynamic-plugin-sdk/src/api/internal-types';
-import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
+import { NamespacedPage } from '@console/shared';
 import { ProjectHelmChartRepositoryType } from '../../../types/helm-types';
 import CreateProjectHelmChartRepository from './CreateProjectHelmChartRepository';
 import { getDefaultResource } from './projecthelmchartrepository-create-utils';
