+++ frontend/packages/helm-plugin/src/components/forms/install-upgrade/HelmInstallUpgradePage.tsx
@@ -13,7 +13,7 @@ import { history, LoadingBox } from '@console/internal/components/utils';
 import { SecretModel } from '@console/internal/models';
 import { k8sGet } from '@console/internal/module/k8s';
 import { prune } from '@console/shared/src/components/dynamic-form/utils';
-import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
+import NamespacedPage from '@console/shared/src/components/NamespacedPage';
 import { EditorType } from '@console/shared/src/components/synced-editor/editor-toggle';
 import {
   HelmActionType,
