+++ frontend/packages/pipelines-plugin/src/components/pac/PacPage.tsx
@@ -5,8 +5,7 @@ import { useAccessReview } from '@console/dynamic-plugin-sdk';
 import { NamespacedPageVariants } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 import { ErrorPage404 } from '@console/internal/components/error';
 import { LoadingBox, AccessDenied } from '@console/internal/components/utils';
-import { useFlag } from '@console/shared/src';
-import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
+import { useFlag, NamespacedPage } from '@console/shared/src';
 import { FLAG_OPENSHIFT_PIPELINE } from '../../const';
 import { PIPELINE_NAMESPACE } from '../pipelines/const';
 import { usePacData } from './hooks/usePacData';
