+++ frontend/packages/pipelines-plugin/src/components/pipelines/PipelineTabbedPage.tsx
@@ -5,9 +5,14 @@ import { RouteComponentProps } from 'react-router-dom';
 import { NamespacedPageVariants } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 import { withStartGuide } from '@console/internal/components/start-guide';
 import { Page } from '@console/internal/components/utils';
-import { MenuAction, MenuActions, MultiTabListPage, useFlag } from '@console/shared';
-import CreateProjectListPage from '@console/shared/src/components/projects/CreateProjectListPage';
-import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
+import {
+  MenuAction,
+  MenuActions,
+  MultiTabListPage,
+  useFlag,
+  CreateProjectListPage,
+  NamespacedPage,
+} from '@console/shared';
 import { FLAG_OPENSHIFT_PIPELINE_AS_CODE } from '../../const';
 import { PipelineModel, RepositoryModel } from '../../models';
 import { usePipelineTechPreviewBadge } from '../../utils/hooks';
