+++ frontend/packages/pipelines-plugin/src/components/tasks/list-page/TasksListsPage.tsx
@@ -5,8 +5,7 @@ import { NamespacedPageVariants } from '@console/dynamic-plugin-sdk/src/api/inte
 import { DefaultPage } from '@console/internal/components/default-resource';
 import { Page } from '@console/internal/components/utils';
 import { referenceForModel } from '@console/internal/module/k8s';
-import { MenuActions, MultiTabListPage } from '@console/shared';
-import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
+import { MenuActions, MultiTabListPage, NamespacedPage } from '@console/shared';
 import { TaskModel, ClusterTaskModel, TaskRunModel } from '../../../models';
 import { usePipelineTechPreviewBadge } from '../../../utils/hooks';
 import TaskRunsListPage from '../../taskruns/list-page/TaskRunsListPage';
