+++ frontend/packages/topology/src/__tests__/TopologyPage.spec.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
 import { useK8sWatchResources } from '@console/internal/components/utils/k8s-watch-hook';
-import { useQueryParams, useUserSettingsCompatibility } from '@console/shared/src';
-import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
+import { useQueryParams, useUserSettingsCompatibility, NamespacedPage } from '@console/shared/src';
 import { TopologyPage } from '../components/page/TopologyPage';
 import { TopologyViewType } from '../topology-types';
 import { usePreferredTopologyView } from '../user-preferences/usePreferredTopologyView';
