+++ frontend/packages/topology/src/components/page/TopologyPage.tsx
@@ -6,10 +6,13 @@ import { match as RMatch } from 'react-router-dom';
 import { NamespacedPageVariants } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 import { withStartGuide } from '@console/internal/components/start-guide';
 import { removeQueryArgument, setQueryArgument } from '@console/internal/components/utils';
-import { useQueryParams, useUserSettingsCompatibility } from '@console/shared';
+import {
+  useQueryParams,
+  useUserSettingsCompatibility,
+  CreateProjectListPage,
+  NamespacedPage,
+} from '@console/shared';
 import { ErrorBoundaryFallbackPage, withFallback } from '@console/shared/src/components/error';
-import CreateProjectListPage from '@console/shared/src/components/projects/CreateProjectListPage';
-import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
 import {
   LAST_TOPOLOGY_OVERVIEW_OPEN_STORAGE_KEY,
   LAST_TOPOLOGY_VIEW_LOCAL_STORAGE_KEY,
